requirements.txt
google-cloud-bigquery == 2.20.0